# api/openapi.yaml
openapi: 3.1.0
info:
  title: Billmesh API
  version: 1.0.0
  description: |
    Billing, subscriptions, credit wallets, metering, payments, and webhook APIs.

    Service clients call `/v1/*` with a bearer token. When the browser BFF is
    enabled, the same protected operations are available below `/api/v1/*`
    with a realm-specific opaque cookie. Unsafe BFF requests additionally
    require the `X-CSRF-Token` returned by that realm's session endpoint.
servers:
  - url: /
    description: Current Billmesh server
...
security:
  - bearerAuth: []
paths:
...
  /v1/accounts:
    post:
      operationId: createAccount
      tags: [Accounts]
      summary: Create or return a billing account for an identity
      x-required-permissions: [billing:write]
      requestBody: { $ref: "#/components/requestBodies/CreateAccount" }
      responses:
        "200": { $ref: "#/components/responses/Account" }
        "201": { $ref: "#/components/responses/Account" }
        default: { $ref: "#/components/responses/Error" }
  /v1/accounts/current:
    get:
      operationId: getCurrentAccount
      tags: [Accounts]
      summary: Get the account linked to the caller
      x-required-permissions: [billing:read]
      responses:
        "200": { $ref: "#/components/responses/Account" }
        default: { $ref: "#/components/responses/Error" }
  /v1/accounts/{id}:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    get:
      operationId: getAccount
      tags: [Accounts]
      summary: Get a billing account
      x-required-permissions: [billing:read]
      responses:
        "200": { $ref: "#/components/responses/Account" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateAccount
      tags: [Accounts]
      summary: Rename a billing account
      x-required-permissions: [billing:write]
      requestBody: { $ref: "#/components/requestBodies/UpdateAccount" }
      responses:
        "200": { $ref: "#/components/responses/AccountSummary" }
        default: { $ref: "#/components/responses/Error" }
  /v1/accounts/{id}/links:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    post:
      operationId: linkAccount
      tags: [Accounts]
      summary: Link another application identity to an account
      x-required-permissions: [billing:write]
      requestBody: { $ref: "#/components/requestBodies/AccountLink" }
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        default: { $ref: "#/components/responses/Error" }
  /v1/account-links:
    post:
      operationId: linkCurrentAccount
      tags: [Accounts]
      summary: Link another application identity to the caller's account
      x-required-permissions: [billing:write]
      requestBody: { $ref: "#/components/requestBodies/AccountLink" }
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        default: { $ref: "#/components/responses/Error" }
  /v1/accounts/{id}/installations:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    post:
      operationId: createInstallation
      tags: [Installations]
      summary: Create or reactivate an application installation
      x-required-permissions: [billing:write]
      requestBody: { $ref: "#/components/requestBodies/Installation" }
      responses:
        "201":
          description: Installation created or reactivated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Installation" }
        default: { $ref: "#/components/responses/Error" }
  /v1/products:
    get:
      operationId: listProducts
      tags: [Catalogue]
      summary: List products
      responses:
        "200":
          description: Products.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createProduct
      tags: [Catalogue]
      summary: Create a product
      x-required-permissions: [billing:admin]
      requestBody: { $ref: "#/components/requestBodies/CreateProduct" }
      responses:
        "201":
          description: Product created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        default: { $ref: "#/components/responses/Error" }
  /v1/plans:
    get:
      operationId: listPlans
      tags: [Catalogue]
      summary: List active plans
      parameters:
        - { $ref: "#/components/parameters/ProductFilter" }
      responses:
        "200":
          description: Plans.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Plan" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createPlan
      tags: [Catalogue]
      summary: Create a plan
      x-required-permissions: [billing:admin]
      requestBody: { $ref: "#/components/requestBodies/CreatePlan" }
      responses:
        "201": { $ref: "#/components/responses/IDWithSlug" }
        default: { $ref: "#/components/responses/Error" }
  /v1/plans/{id}:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    patch:
      operationId: updatePlan
      tags: [Catalogue]
      summary: Update a plan
      x-required-permissions: [billing:admin]
      requestBody: { $ref: "#/components/requestBodies/UpdatePlan" }
      responses:
        "200": { $ref: "#/components/responses/IDOnly" }
        default: { $ref: "#/components/responses/Error" }
  /v1/credit-packs:
    post:
      operationId: createCreditPack
      tags: [Catalogue]
      summary: Create a prepaid credit pack
      x-required-permissions: [billing:admin]
      requestBody: { $ref: "#/components/requestBodies/CreateCreditPack" }
      responses:
        "201": { $ref: "#/components/responses/IDOnly" }
        default: { $ref: "#/components/responses/Error" }
  /v1/subscriptions:
    post:
      operationId: createSubscription
      tags: [Subscriptions]
      summary: Subscribe an account to a plan
      x-required-permissions: [billing:write]
      requestBody: { $ref: "#/components/requestBodies/CreateSubscription" }
      responses:
        "201":
          description: Subscription created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subscription" }
        default: { $ref: "#/components/responses/Error" }
  /v1/subscriptions/current:
    get:
      operationId: getCurrentSubscription
      tags: [Subscriptions]
      summary: Get the caller's current subscription
      x-required-permissions: [billing:read]
      parameters:
        - { $ref: "#/components/parameters/ProductFilter" }
      responses:
        "200":
          description: Current subscription.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subscription" }
        default: { $ref: "#/components/responses/Error" }
  /v1/subscriptions/{id}/change-plan:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    post:
      operationId: changeSubscriptionPlan
      tags: [Subscriptions]
      summary: Change a subscription's plan
      x-required-permissions: [billing:write]
      requestBody: { $ref: "#/components/requestBodies/ChangePlan" }
      responses:
        "200": { $ref: "#/components/responses/SubscriptionPlan" }
        default: { $ref: "#/components/responses/Error" }
  /v1/subscriptions/{id}/cancel:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    post:
      operationId: cancelSubscription
      tags: [Subscriptions]
      summary: Cancel a subscription immediately or at period end
      x-required-permissions: [billing:write]
      requestBody: { $ref: "#/components/requestBodies/SubscriptionStatus" }
      responses:
        "200": { $ref: "#/components/responses/SubscriptionStatus" }
        default: { $ref: "#/components/responses/Error" }
  /v1/subscriptions/current/cancel:
    post:
      operationId: cancelCurrentSubscription
      tags: [Subscriptions]
      summary: Immediately cancel the caller's current application subscription
      x-required-permissions: [billing:write]
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        default: { $ref: "#/components/responses/Error" }
  /v1/subscriptions/{id}/reactivate:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    post:
      operationId: reactivateSubscription
      tags: [Subscriptions]
      summary: Reactivate a subscription
      x-required-permissions: [billing:write]
      requestBody: { $ref: "#/components/requestBodies/SubscriptionStatus" }
      responses:
        "200": { $ref: "#/components/responses/SubscriptionStatus" }
        default: { $ref: "#/components/responses/Error" }
  /v1/subscriptions/{id}/renew:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    post:
      operationId: renewSubscription
      tags: [Subscriptions]
      summary: Renew a subscription idempotently
      x-required-permissions: [billing:write]
      requestBody: { $ref: "#/components/requestBodies/RenewSubscription" }
      responses:
        "200":
          description: Renewal result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Renewal" }
        default: { $ref: "#/components/responses/Error" }
  /v1/entitlements:
    get:
      operationId: getEntitlements
      tags: [Entitlements]
      summary: Get the effective entitlements for a product
      x-required-permissions: [billing:read]
      parameters:
        - { $ref: "#/components/parameters/ProductFilter" }
      responses:
        "200":
          description: Subscription status and entitlements.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Entitlements" }
        default: { $ref: "#/components/responses/Error" }
  /v1/entitlements/check:
    get:
      operationId: checkEntitlement
      tags: [Entitlements]
      summary: Check whether a feature is enabled
      x-required-permissions: [billing:read]
      parameters:
        - name: feature
          in: query
          required: true
          schema: { type: string }
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        default: { $ref: "#/components/responses/Error" }
  /v1/wallets:
    post:
      operationId: createWallet
      tags: [Wallets]
      summary: Create or return an account's product wallet
      x-required-permissions: [billing:write]
      requestBody: { $ref: "#/components/requestBodies/CreateWallet" }
      responses:
        "201": { $ref: "#/components/responses/Wallet" }
        default: { $ref: "#/components/responses/Error" }
  /v1/wallets/{id}:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    get:
      operationId: getWallet
      tags: [Wallets]
      summary: Get wallet balances
      x-required-permissions: [billing:read]
      responses:
        "200": { $ref: "#/components/responses/Wallet" }
        default: { $ref: "#/components/responses/Error" }
  /v1/wallets/{id}/ledger:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    get:
      operationId: listWalletLedger
      tags: [Wallets]
      summary: List a wallet's credit ledger
      x-required-permissions: [billing:read]
      parameters:
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200":
          description: Ledger entries in creation order.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LedgerEntry" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token carrying Billmesh permissions and application context.
    consoleBrowserSession:
      type: apiKey
      in: cookie
      name: __Host-billmesh-console-session
      description: Opaque Console session cookie owned by the Billmesh API origin.
    adminBrowserSession:
      type: apiKey
      in: cookie
      name: __Host-billmesh-admin-session
      description: Opaque Admin session cookie owned by the Billmesh API origin.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required with the browser session cookie for unsafe BFF requests.
    razorpaySignature:
      type: apiKey
      in: header
      name: X-Razorpay-Signature
      description: HMAC-SHA256 signature supplied by Razorpay.
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UUID" }
...
  responses:
...
    Account:
      description: Billing account.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Account" }
...